+++ lib/routes/stats/insights_page.dart
@@ -1,5 +1,6 @@
 import "package:flow/l10n/extensions.dart";
 import "package:flow/widgets/general/frame.dart";
+import "package:flow/widgets/home/stats/bento/budget_tile.dart";
 import "package:flow/widgets/home/stats/bento/calendar_tile.dart";
 import "package:flow/widgets/home/stats/bento/map_tile.dart";
 import "package:flow/widgets/home/stats/bento/net_worth_tile.dart";
@@ -38,6 +39,8 @@ class InsightsPage extends StatelessWidget {
                   const WrappedTile(),
                   const SizedBox(height: 12.0),
                   const NetWorthTile(),
+                  const SizedBox(height: 12.0),
+                  const BudgetTile(),
                   const SizedBox(height: 12.0),
                   const Row(
                     spacing: 12.0,